create DriverSearchRequestActivity and modified DriverMainActivity
