+++ app/src/main/AndroidManifest.xml
@@ -18,10 +18,10 @@
         <activity android:name=".view.SignupActivity" />
         <activity
             android:name=".view.DriverMainActivity"
-            android:label="@string/title_activity_driver_main"
             android:theme="@style/AppTheme.NoActionBar" />
         <activity android:name=".view.EditUserProfileActivity" />
-        <activity android:name=".view.DriverBrowseRequestActivity"></activity>
+        <activity android:name=".view.DriverBrowseRequestActivity" />
+        <activity android:name=".view.DriverSearchRequestActivity"></activity>
     </application>
 
 </manifest>
