+++ app/src/main/res/layout/content_driver_main.xml
@@ -13,35 +13,6 @@
     tools:context="ca.ualberta.cs.unter.view.DriverMainActivity"
     tools:showIn="@layout/app_bar_driver_main">
 
-    <EditText android:id="@+id/editText_searchRequest_DriverMainActivity"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:hint="@string/search_request"
-        android:inputType="textNoSuggestions|textVisiblePassword"
-        android:paddingTop="30dp"
-        android:maxLines="1"/>
-
-    <Spinner
-        android:id="@+id/spinner_searchOption_DriverMainActivity"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:paddingTop="5dp"/>
-
-    <TextView
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="@string/search_result"
-        android:paddingTop="20dp"
-        android:textSize="20sp"
-        android:textStyle="bold"/>
-
-    <ListView
-        android:id="@+id/listView_searchList_DriverMainActivity"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:scrollbars="vertical"
-        android:clickable="true" >
-    </ListView>
 
 </LinearLayout>
